removed old code and renamed committees_test.py to committees.py
